+++ changelog.md
@@ -1,3 +1,29 @@
+## 9.0
+
+This is a full rewrite.
+
+- Promise API instead of callbacks.
+- Accept pattern as string or array of strings.
+- Hybrid module distribution.
+- Full TypeScript support.
+- Removed `root` option and mounting behavior.
+- Removed `stat` option.
+- Simplified `cwd` behavior.  Now it simply serves as the initial
+  argument to `fs.readdir`.
+- Removed all stat calls, in favor of using `withFileTypes:true`
+  with `fs.readdir()`.
+- Consolidated all caching to a single object that only caches
+  directory entries and `readdir` errors.
+- Removed EventEmitter behavior from exported `Glob` class.
+- Consolidated sync and async `Glob` class behavior into a single
+  class with `process()` and `processSync()` methods.
+- Removed `silent` option.  Any readdir errors are simply treated as
+  "the directory could not be read", and it is treated as a
+  normal file entry instead, like shells do.
+- Removed `fs` option.  This module only operates on the real
+  filesystem.
+- Only support node 16 and higher.
+
 ## 8.1
 
 - Add `windowsPathsNoEscape` option